Moved launch functionality into __init__.py to avoid circular import issue.
