+++ src/protocol_proxy/protocol/bacnet/bacnet_proxy.py
@@ -7,11 +7,9 @@ import time
 import traceback
 import csv
 
-from argparse import ArgumentParser
 from dataclasses import dataclass
 from datetime import datetime
 from functools import partial
-from typing import Type
 
 from bacpypes3.basetypes import PropertyReference
 from bacpypes3.lib.batchread import BatchRead, DeviceAddressObjectPropertyReference
@@ -1162,23 +1160,6 @@ class BACnetProxy(AsyncioProtocolProxy):
             _log.error(f"General error during Who-Is: {e_gen}")
             return []
 
-
-async def run_proxy(local_device_address, **kwargs):
-    bp = BACnetProxy(local_device_address, **kwargs)
-    await bp.start()
-
-
-def launch_bacnet(parser: ArgumentParser) -> tuple[ArgumentParser, Type[AsyncioProtocolProxy]]:
-    parser.add_argument('--local-device-address', type=str, required=True,
-                        help='Address on the local machine of this BACnet Proxy.')
-    parser.add_argument('--bacnet-network', type=int, default=0,
-                        help='The BACnet port as an offset from 47808.')
-    parser.add_argument('--vendor-id', type=int, default=999,
-                        help='The BACnet vendor ID to use for the local device of this BACnet Proxy.')
-    parser.add_argument('--object-name', type=str, default='VOLTTRON BACnet Proxy',
-                        help='The name of the local device for this BACnet Proxy.')
-    return parser, run_proxy
-
-
 if __name__ == '__main__':
+    from . import launch_bacnet
     sys.exit(launch(launch_bacnet))
